refactor: splitting stores out into separate files
